PR conversation fixes
