+++ src/Lykke.Service.BlockchainApi.Sdk/Constants.cs
