+++ src/Lykke.Service.BlockchainApi.Sdk/Controllers/AssetsController.cs
@@ -3,20 +3,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Lykke.Service.BlockchainApi.Contract;
 using Lykke.Service.BlockchainApi.Contract.Assets;
-using Lykke.Service.BlockchainApi.Sdk.Domain;
+using Lykke.Service.BlockchainApi.Sdk.Domain.Assets;
 using Lykke.Service.BlockchainApi.Sdk.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Lykke.Service.BlockchainApi.Sdk.Controllers
 {
-    public class CreateAssetRequest
-    {
-        [Required]     public string AssetId  { get; set; }
-                       public string Address  { get; set; }
-        [Required]     public string Name     { get; set; }
-        [Range(0, 28)] public int    Accuracy { get; set; }
-    }
-
     [ApiController]
     [Route("/api/assets")]
     public class AssetsController : ControllerBase
@@ -26,7 +18,9 @@ namespace Lykke.Service.BlockchainApi.Sdk.Controllers
         public AssetsController(AssetRepository assets) => _assets = assets;
         
         [HttpGet]
-        public async Task<ActionResult<PaginationResponse<AssetContract>>> Get([Range(1, int.MaxValue)] int take, [AzureContinuation] string continuation)
+        public async Task<ActionResult<PaginationResponse<AssetContract>>> Get(
+            [Range(1, int.MaxValue)] int take, 
+            [AzureContinuation] string continuation)
         {
             var chunk = await _assets.GetAsync(take, continuation);
